adding dedicated data module
