+++ aframe/pipelines/configs/train.yaml
@@ -7,29 +7,30 @@ model:
   arch:
     class_path: aframe.architectures.ResNet
     init_args:
-      # num_ifos:
       layers: [3, 4, 6, 3]
       norm_groups: 16
-
-  # data/preprocessing args
+  metric:
+    class_path: torchmetrics.classification.BinaryAUROC
+    init_args:
+      max_fpr: 1e-3
+data:
+  # loading args
   # data_dir:
   # ifos:
+
+  # preprocessing args
+  batch_size: 384
   kernel_length: 1.5
   psd_length: 8
   fduration: 1
   highpass: 32
   fftlength: null
 
-  # optimization args
-  batch_size: 384
-  max_lr: 0.000585
-  lr_ramp_epochs: 23
+  # augmentation args
+  waveform_prob: 0.277
   snr_thresh: 4
   max_snr: 100
   snr_alpha: 3
-
-  # augmentation args
-  waveform_prob: 0.277
   swap_frac: 0.014
   mute_frac: 0.055
   trigger_pad: -0.75
@@ -39,6 +40,10 @@ model:
   valid_stride: 0.5
   num_valid_views: 4
   valid_livetime: 43200
+optimizer:
+  lr: 0.000585
+lr_scheduler:
+  pct_start: 0.115
 trainer:
   logger:
     - class_path: lightning.pytorch.loggers.CSVLogger
